[cleanup] Eliminate horribly dynamic untyped Modal.vue
