+++ src/components/Modals/EditSemester.vue
@@ -95,63 +95,6 @@ export default Vue.extend({
 <style lang="scss">
 @import '@/assets/scss/_variables.scss';
 
-.modal {
-  padding: 1rem;
-
-  &-content {
-    background: $white;
-    border-radius: 9px;
-    margin-left: auto;
-    margin-right: auto;
-    padding: 1rem;
-  }
-
-  &-body {
-    padding: 0;
-  }
-
-  &-top {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 0.5rem;
-  }
-
-  &-exit {
-    width: 10.5px;
-    height: 10.5px;
-    cursor: pointer;
-  }
-
-  &-title {
-    font-weight: bold;
-    font-size: 16px;
-    line-height: 20px;
-    color: #3d3d3d;
-  }
-
-  &-buttonWrapper {
-    margin-top: 1rem;
-    display: flex;
-    justify-content: flex-end;
-  }
-
-  &-button {
-    width: 4.75rem;
-    color: #5b676d;
-    border-radius: 3px;
-    border: 1px solid #3d3d3d;
-    background-color: $white;
-    display: flex;
-    justify-content: center;
-
-    &--add {
-      color: $white;
-      background-color: $sangBlue;
-      margin-left: 0.5rem;
-      border: none;
-    }
-  }
-}
 .editSemesterModal {
   padding: 1rem;
